refactor: download npm
