+++ lib/download/npm.js
@@ -3,7 +3,7 @@
  * MIT Licensed
  *
  * Authors:
- *   fengmk2 <m@fengmk2.com> (http://fengmk2.com)
+ *   dead_horse <dead_horse@qq.com>
  */
 
 'use strict';
@@ -12,18 +12,54 @@
  * Module dependencies.
  */
 
-const debug = require('debug')('npminstall:download');
+const debug = require('debug')('npminstall:download:npm');
 const bytes = require('bytes');
 const fs = require('mz/fs');
 const path = require('path');
 const crypto = require('crypto');
 const tar = require('tar');
+const url = require('url');
 const zlib = require('zlib');
 const destroy = require('destroy');
-const get = require('./get');
-const utils = require('./utils');
+const utility = require('utility');
+const get = require('../get');
+const utils = require('../utils');
 
-module.exports = function* download(pkg, options) {
+module.exports = function* (pkg, options) {
+  // get npm package info
+  const pkgUrl = getPackageNpmUri(pkg, options);
+  const result = yield get(pkgUrl, {
+    dataType: 'json',
+    timeout: options.timeout,
+    followRedirect: true,
+    gzip: true,
+  });
+  const realPkg = result.data;
+  options.totalJSONSize += result.res.size;
+  debug('[%s@%s] status %s, real version: %s, headers: %j, responseSize: %s',
+    pkg.name, pkg.version, result.status, realPkg.version, result.headers, result.res.size);
+
+  // download tarball and unzip
+  const info = yield download(realPkg, options);
+  info.package = realPkg;
+  return info;
+};
+
+function getPackageNpmUri(pkg, options) {
+  let name = pkg.name;
+  if (name[0] === '@') {
+    // dont encodeURIComponent @ char, it will be 405
+    // https://registry.npmjs.org/%40rstacruz%2Ftap-spec/%3E%3D4.1.1
+    name = '@' + utility.encodeURIComponent(name.substring(1));
+  } else {
+    name = utility.encodeURIComponent(name);
+  }
+  const pkgUrl = url.resolve(options.registry, name + '/' + utility.encodeURIComponent(pkg.spec));
+  debug('[%s@%s] GET %j', pkg.name, pkg.spec, pkgUrl);
+  return pkgUrl;
+}
+
+function* download(pkg, options) {
   const ungzipDir = path.join(options.storeDir, pkg.name, pkg.version);
   const donefile = path.join(ungzipDir, '.tnpminstall.done');
 
@@ -67,7 +103,7 @@ module.exports = function* download(pkg, options) {
     exists: false,
     dir: ungzipDir,
   };
-};
+}
 
 function* getTarballStream(pkg, options) {
   if (!options.cacheDir) {
