+++ lib/install.js
@@ -14,13 +14,10 @@
 
 const debug = require('debug')('npminstall:install');
 const path = require('path');
-const url = require('url');
-const utility = require('utility');
 const npa = require('npm-package-arg');
 const chalk = require('chalk');
 const semver = require('semver');
 const fs = require('mz/fs');
-const get = require('./get');
 const download = require('./download');
 const utils = require('./utils');
 const postinstall = require('./postinstall');
@@ -53,6 +50,7 @@ function* _install(parentDir, pkg, options) {
   debug('install %s@%s in %s', pkg.name, pkg.version, parentDir);
 
   const p = npa(pkg.name + '@' + pkg.version);
+
   if (!p || p.type === 'remote' || p.type === 'git' || p.type === 'hosted') {
     console.error(chalk.red(`[${pkg.name}@${pkg.version}] is not supported yet, type: ${p.type}, parentDir: ${parentDir}`));
     return '';
@@ -87,21 +85,8 @@ function* _install(parentDir, pkg, options) {
     }
   }
 
-  // get npm package info
-  const pkgUrl = getPackageNpmUri(pkg, options);
-  const result = yield get(pkgUrl, {
-    dataType: 'json',
-    timeout: options.timeout,
-    followRedirect: true,
-    gzip: true,
-  });
-  const realPkg = result.data;
-  options.totalJSONSize += result.res.size;
-  debug('[%s@%s] status %s, real version: %s, headers: %j, responseSize: %s',
-    pkg.name, pkg.version, result.status, realPkg.version, result.headers, result.res.size);
-
-  // download tarball and unzip
-  const info = yield download(realPkg, options);
+  const info = yield download(p, options);
+  const realPkg = info.package;
   const realPkgDir = info.dir;
   options.cache[key] = {
     package: realPkg,
@@ -191,18 +176,4 @@ function* bundleBin(name, parentDir) {
   const pkgfile = path.join(pkgDir, 'package.json');
   const pkg = yield utils.readJSON(pkgfile);
   yield bin(parentDir, pkg, pkgDir);
-}
-
-function getPackageNpmUri(pkg, options) {
-  let name = pkg.name;
-  if (name[0] === '@') {
-    // dont encodeURIComponent @ char, it will be 405
-    // https://registry.npmjs.org/%40rstacruz%2Ftap-spec/%3E%3D4.1.1
-    name = '@' + utility.encodeURIComponent(name.substring(1));
-  } else {
-    name = utility.encodeURIComponent(name);
-  }
-  const pkgUrl = url.resolve(options.registry, name + '/' + utility.encodeURIComponent(pkg.version));
-  debug('[%s@%s] GET %j', pkg.name, pkg.version, pkgUrl);
-  return pkgUrl;
 }
